Use #repr(..) to prevent unnecessary casts

+++ examples/plugwise_test.rs
@@ -241,18 +241,18 @@ impl ReqSwitch {
     }
 }
 
-const EMPTY: isize = 0x0000;
-const REQ_INITIALIZE: isize = 0x000A;
-const RES_INITIALIZE: isize = 0x0011;
-const REQ_INFO: isize = 0x0023;
-const RES_INFO: isize = 0x0024;
-const REQ_SWITCH: isize = 0x0017;
-// FIXME: const RES_CALIBRATION: isize = 0x0027;
+const EMPTY: u16 = 0x0000;
+const REQ_INITIALIZE: u16 = 0x000A;
+const RES_INITIALIZE: u16 = 0x0011;
+const REQ_INFO: u16 = 0x0023;
+const RES_INFO: u16 = 0x0024;
+const REQ_SWITCH: u16 = 0x0017;
+// FIXME: const RES_CALIBRATION: u16 = 0x0027;
 //  - double: gain_a
 //  - double: gain_b
 //  - double: off_total
 //  - double: off_noise
-// FIXME: const RES_CLOCK_INFO: isize = 0x003F;
+// FIXME: const RES_CLOCK_INFO: u16 = 0x003F;
 //  - time:
 //      - u8: hour
 //      - u8: minute
@@ -260,14 +260,14 @@ const REQ_SWITCH: isize = 0x0017;
 //  - u8: day of week
 //  - u8: unknown
 //  - u16: unknown
-// FIXME: const RES_POWER_USE: isize = 0x0013;
+// FIXME: const RES_POWER_USE: u16 = 0x0013;
 //  - u16: pulse 1s
 //  - u16: pulse 8s
 //  - u32: pulse hour
 //  - u16: unknown
 //  - u16: unknown
 //  - u16: unknown
-// FIXME: const RES_POWER_BUFFER: isize = 0x0049;
+// FIXME: const RES_POWER_BUFFER: u16 = 0x0049;
 //  - u32: datetime1 (see DateTime)
 //  - u32: pulses1
 //  - u32: datetime2 (see DateTime)
@@ -277,17 +277,18 @@ const REQ_SWITCH: isize = 0x0017;
 //  - u32: datetime4 (see DateTime)
 //  - u32: pulses4
 //  - u32: logaddr
-// FIXME: const REQ_POWER_USE: isize = 0x0012;
-// FIXME: const REQ_CLOCK_INFO: isize = 0x003E;
-// FIXME: const REQ_CLOCK_SET: isize = 0x0016;
+// FIXME: const REQ_POWER_USE: u16 = 0x0012;
+// FIXME: const REQ_CLOCK_INFO: u16 = 0x003E;
+// FIXME: const REQ_CLOCK_SET: u16 = 0x0016;
 //  - u32: datetime (see DateTime)
 //  - u32: logaddr
 //  - u24: time (see 003F)
 //  - u8: day of week
-// FIXME: const REQ_CALIBRATION: isize = 0x0026;
-// FIXME: const REQ_POWER_BUFFER: isize = 0x0048;
+// FIXME: const REQ_CALIBRATION: u16 = 0x0026;
+// FIXME: const REQ_POWER_BUFFER: u16 = 0x0048;
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
+#[repr(u16)]
 enum MessageId {
     Empty = EMPTY,
     ReqInitialize = REQ_INITIALIZE,
@@ -299,7 +300,7 @@ enum MessageId {
 
 impl MessageId {
     fn new(id: u16) -> MessageId {
-        match id as isize {
+        match id {
             EMPTY => MessageId::Empty,
             REQ_INITIALIZE => MessageId::ReqInitialize,
             RES_INITIALIZE => MessageId::ResInitialize,
